+++ apps/server/src/api/products/(products)/kvstore/models/kv.model.ts
@@ -1,4 +1,4 @@
-import { kvstoreMongoDbConn } from "../../../../lib/db-connect"
+import { kvstoreMongoDbConn } from "../../../../../lib/db-connect"
 import { KvSchema } from "../schemas/kv.schema"
 
 export const KvModel = kvstoreMongoDbConn.model("kv", KvSchema)