+++ apps/server/src/api/products/(products)/ledgerscan/models/transaction.model.ts
@@ -1,4 +1,4 @@
-import { ledgerscanMongoDbConn } from "../../../../lib/db-connect"
+import { ledgerscanMongoDbConn } from "../../../../../lib/db-connect"
 import { TransactionSchema } from "../schemas/transaction.schema"
 
 export const TransactionModel = ledgerscanMongoDbConn.model("transaction", TransactionSchema)