+++ apps/server/src/api/products/(products)/nftstudio/models/transaction.model.ts
@@ -1,4 +1,4 @@
-import { nftstudioMongoDbConn } from "../../../../lib/db-connect"
+import { nftstudioMongoDbConn } from "../../../../../lib/db-connect"
 import { TransactionSchema } from "../schemas/transaction.schema"
 
 export const TransactionModel = nftstudioMongoDbConn.model("transaction", TransactionSchema)