+++ apps/server/src/api/products/(products)/nftstudio/nftstudio.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Post, BadRequestException, Get, Body } from "@nestjs/common"
 import { NftstudioService } from "./nftstudio.service"
-import { statusMessages } from "../../../constants/status-messages"
+import { statusMessages } from "../../../../constants/status-messages"
 import { TokenAuthorizer, TokenAuthorizerResponse } from "src/authorization/token-authorizer.decorator"
 import { CredentialAuthorizer, CredentialAuthorizerResponse } from "src/authorization/credential-authorizer.decorator"